fix: a huge optimization not to rerender Cards on Map + updated linter target paths
