+++ client/src/components/Map/index.tsx
@@ -1,9 +1,6 @@
-import React, { useMemo, useState, useEffect } from 'react';
+import React, { useMemo, useState, useEffect, useCallback } from 'react';
 import * as d3 from 'd3';
 import _ from 'lodash';
-
-// import { LinkFromEndpointToConnector } from './components/LinkFromEndpointToConnector';
-// import { MapGraph } from './entities';
 
 import { useZoom } from '~/ui/hooks/useZoom';
 import { setCSSVars } from '~/ui';
@@ -12,11 +9,15 @@ import { Endpoint } from '~/domain/endpoint';
 import { dummy as geom } from '~/domain/geometry';
 import { sizes } from '~/ui/vars';
 import { useStore } from '~/store/hooks';
+import { XYWH } from '~/domain/geometry';
+import { Placement, PlacementEntry } from '~/domain/layout';
 
 import {
   EndpointCardBackplate,
   EndpointCardContent,
 } from '~/components/EndpointCard';
+
+import { NamespaceBackplate } from './NamespaceBackplate';
 
 import css from './styles.scss';
 
@@ -26,73 +27,55 @@ export interface Props {
   readonly namespace: string | undefined;
 }
 
+export interface MapElementsProps {
+  namespace: string | undefined;
+}
+
+const MapElements = React.memo(function MapElements(props: MapElementsProps) {
+  const { layout } = useStore();
+  const { namespace } = props;
+  const [nsXYWH, setNsXYWH] = useState(geom.xywh());
+
+  const placement = useMemo(() => layout.placement, [layout]);
+
+  const updateNamespaceLayer = useCallback(() => {
+    const nsBBox = layout.cardsBBox.addMargin(sizes.endpointHPadding / 2);
+    setNsXYWH(nsBBox);
+  }, []);
+
+  return (
+    <>
+      <NamespaceBackplate namespace={namespace} xywh={nsXYWH} />
+
+      {placement.map(card => (
+        <EndpointCardBackplate
+          key={card.endpoint.id!}
+          coords={card.geometry}
+          endpoint={card.endpoint}
+          onHeightChange={h => updateNamespaceLayer()}
+        />
+      ))}
+
+      {placement.map(card => (
+        <EndpointCardContent
+          key={card.endpoint.id!}
+          coords={card.geometry}
+          endpoint={card.endpoint}
+          onHeightChange={h => updateNamespaceLayer()}
+        />
+      ))}
+    </>
+  );
+});
+
 const MapComponent = (props: Props) => {
   const ref = React.useRef<SVGSVGElement>(null);
   const zoomProps = useZoom(ref, { tx: sizes.endpointHPadding });
-  const [nsXYWH, setNsXYWH] = useState(geom.xywh());
-  const { layout } = useStore();
-
-  const updateNamespaceLayer = () => {
-    const nsBBox = layout.cardsBBox.addMargin(sizes.endpointHPadding / 2);
-    setNsXYWH(nsBBox);
-  };
-
-  const onCardHeightChange = (_: any, newHeight: number) => {
-    updateNamespaceLayer();
-  };
-
-  useEffect(() => {
-    updateNamespaceLayer();
-  }, [zoomProps]);
-
-  const placement = layout.placement;
-
-  const NamespaceBackplate = props.namespace != null && (
-    <g
-      className={css.namespaceBackplate}
-      transform={`translate(${nsXYWH.x}, ${nsXYWH.y})`}
-    >
-      <rect width={nsXYWH.w} height={nsXYWH.h} rx={10} ry={10} />
-      <text
-        x={sizes.endpointHPadding / 2 + sizes.endpointShadowSize}
-        y={sizes.endpointHPadding / 3}
-      >
-        {props.namespace}
-      </text>
-    </g>
-  );
 
   return (
     <svg ref={ref} className={css.wrapper}>
       <g transform={zoomProps ? zoomProps.toString() : ''}>
-        {NamespaceBackplate}
-
-        {placement.map(props => (
-          <EndpointCardBackplate
-            key={props.endpoint.id!}
-            coords={props.geometry}
-            endpoint={props.endpoint}
-            onHeightChange={h => onCardHeightChange(props, h)}
-          />
-        ))}
-
-        {placement.map(props => (
-          <EndpointCardContent
-            key={props.endpoint.id!}
-            coords={props.geometry}
-            endpoint={props.endpoint}
-            onHeightChange={h => onCardHeightChange(props, h)}
-          />
-        ))}
-
-        {/*{layout.mapEachLink(props => (
-          <LinkFromEndpointToConnector
-            {...props}
-            key={props.srcEndpoint.hash + props.dstEndpoint.hash}
-            connectorX={props.dstEndpointX - 30}
-            connectorY={props.dstEndpointY + 30}
-          />
-        ))}*/}
+        <MapElements namespace={props.namespace} />
       </g>
     </svg>
   );
