+++ client/src/domain/layout.ts
@@ -1,0 +1,9 @@
+import { XYWH } from './geometry';
+import { Endpoint } from './endpoint';
+
+export interface PlacementEntry {
+  endpoint: Endpoint;
+  geometry: XYWH;
+}
+
+export type Placement = Array<PlacementEntry>;
